% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_component}
\alias{check_component}
\title{Valid component check}
\usage{
check_component(component)
}
\arguments{
\item{component}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided component(s) to see if they are available in the summary statistics supplied. Takes certain abbreviations/alternate spellings.
Accepted values are:
b, beh, behavior, behaviour, behaviors, behaviours
m, mod, modifier, modifiers
i, ident, identity, identities
s, set, setting, settings
}
