% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{check_data_cols}
\alias{check_data_cols}
\title{Check data columns}
\usage{
check_data_cols(data, program, group = "none", stat = "mean")
}
\arguments{
\item{data}{dataframe to check}

\item{program}{for bayesact or for interact (bayesact can only have three EPA columns, interact can have 3 or 6... could print a warning if 6 are passed to bayesact)}

\item{group}{name of the grouping column (for interact only)}

\item{stat}{statistic type wanted (mean, sd, cov)}
}
\value{
logical
}
\description{
Checks that a dataframe to be reformatted for bayesact or interact has the necessary columns -- term, E, P, A, and sd or cov if needed (bayesact only).
}
