% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{check_data_frame}
\alias{check_data_frame}
\title{check_data_frame}
\usage{
check_data_frame(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
logical
}
\description{
Verify that it's either a data frame object or a tibble
}
