% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_dataset}
\alias{check_dataset}
\title{Valid dataset check}
\usage{
check_dataset(dataset)
}
\arguments{
\item{dataset}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided dataset(s) to see if they are available in the summary statistics supplied.
}
