% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_datatype}
\alias{check_datatype}
\title{Check data type}
\usage{
check_datatype(datatype)
}
\arguments{
\item{datatype}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Check data type
}
\keyword{internal}
