% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_group}
\alias{check_group}
\title{Valid group check}
\usage{
check_group(group)
}
\arguments{
\item{group}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided component(s) to see if they are available in the summary statistics supplied. Takes certain abbreviations/alternate spellings.
Accepted values are:
m, male, man
f, female, woman
a, av, all
}
\keyword{internal}
