% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{check_group_col}
\alias{check_group_col}
\title{Checks validity of grouping column: if not "none", must have just two unique values.}
\usage{
check_group_col(data, group)
}
\arguments{
\item{data}{dataframe to check}

\item{group}{name of grouping column}
}
\value{
logical
}
\description{
Checks validity of grouping column: if not "none", must have just two unique values.
}
