% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_institutions}
\alias{check_institutions}
\title{Check that institutions are valid}
\usage{
check_institutions(institutions)
}
\arguments{
\item{institutions}{list of institutions}
}
\value{
logical
}
\description{
Check that institutions are valid
}
