% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{check_key}
\alias{check_key}
\title{Valid key check}
\usage{
check_key(key)
}
\arguments{
\item{key}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided key(s) to see if they are available in the summary statistics supplied.
}
