% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{check_no_duplicate_terms}
\alias{check_no_duplicate_terms}
\title{check_no_duplicate_terms}
\usage{
check_no_duplicate_terms(terms, group)
}
\arguments{
\item{terms}{the term list}

\item{group}{name of the grouping variable, or "none"}
}
\value{
logical indicating success
}
\description{
Check whether there are duplicate terms and throw the appropriate error or warning.
}
\keyword{internal}
