% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_stat}
\alias{check_stat}
\title{Valid stat check}
\usage{
check_stat(
  stat,
  stats_avail = c("mean", "m", "sd", "standard deviation", "s", "cov", "covar",
    "covariance", "c", "n", "number")
)
}
\arguments{
\item{stat}{string}

\item{stats_avail}{abbreviations allowed}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided stat(s) to see if they are valid designators. Takes certain abbreviations/alternate spellings.
Accepted values are:
mean, m
sd, standard deviation, s
cov, covar, covariance, c
}
\keyword{internal}
