% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_valid}
\alias{check_valid}
\title{check_valid}
\usage{
check_valid(value, allowed)
}
\arguments{
\item{value}{string or list}

\item{allowed}{string or list}
}
\value{
logical
}
\description{
check_valid
}
