% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china1999_identities_av_dict}
\alias{china1999_identities_av_dict}
\title{China 1999 identities (gender = average)}
\format{
A data frame with 582 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institution codes for this term}
...
}
}
\source{
\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html}
}
\usage{
china1999_identities_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
