% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china2000_impressionabo_m_eqn}
\alias{china2000_impressionabo_m_eqn}
\title{China 2000 impression change actor-behavior-object (gender = male)}
\format{
A data frame with 12 rows (terms) and 10 variables:
\describe{
\item{V1}{term}
\item{V2}{equation}
\item{V3}{equation}
\item{V4}{equation}
\item{V5}{equation}
\item{V6}{equation}
\item{V7}{equation}
\item{V8}{equation}
\item{V9}{equation}
\item{V10}{equation}
...
}
}
\source{
Interact (May 2021)
}
\usage{
china2000_impressionabo_m_eqn
}
\description{
These values are reported as given in Interact. If gender is listed as average, this indicates that for the given study and subset, the values provided in Interact for male and female are identical.
}
\details{
For more information about available equation files, call eqn_info()
}
\keyword{datasets}
