% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{create_instcode}
\alias{create_instcode}
\title{Creates institution code strings from logicals indicating category membership}
\usage{
create_instcode(
  component,
  male = F,
  female = F,
  overt = F,
  surmised = F,
  place = F,
  time = F,
  lay = F,
  business = F,
  law = F,
  politics = F,
  academe = F,
  medicine = F,
  religion = F,
  family = F,
  sexual = F,
  monadic = F,
  group = F,
  corporal = F,
  adjective = F,
  adverb = F,
  emotion = F,
  trait = F,
  status = F,
  feature = F,
  emotion_spiral = F
)
}
\arguments{
\item{component}{whether the term is an identity, behavior, component, or setting}

\item{male}{logical identities (whether identities can typically be applied to men)}

\item{female}{logical identities (whether identities can typically be applied to women)}

\item{overt}{logical behaviors}

\item{surmised}{logical behaviors}

\item{place}{logical settings}

\item{time}{logical settings}

\item{lay}{logical identities, behaviors, settings}

\item{business}{logical identities, behaviors, settings}

\item{law}{logical identities, behaviors, settings}

\item{politics}{logical identities, behaviors, settings}

\item{academe}{logical identities, behaviors, settings}

\item{medicine}{logical identities, behaviors, settings}

\item{religion}{logical identities, behaviors, settings}

\item{family}{logical identities, behaviors, settings}

\item{sexual}{logical identities, behaviors, settings}

\item{monadic}{logical identities, behaviors, settings}

\item{group}{logical identities, behaviors, settings}

\item{corporal}{logical identities, behaviors, settings}

\item{adjective}{logical modifiers}

\item{adverb}{logical modifiers}

\item{emotion}{logical modifiers}

\item{trait}{logical modifiers}

\item{status}{logical modifiers}

\item{feature}{logical modifiers}

\item{emotion_spiral}{logical modifiers}
}
\value{
character string institution code
}
\description{
This function returns institution codes in string format that is properly formatted
for import to Interact and consistent with typical ACT data practices. Relevant
arguments depend on the component. Arguments that do not apply to the given
component will be ignored.
}
\details{
Identities: male, female, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal
Behaviors: overt, surmised, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal
Modifiers: adjective, adverb, emotion, trait, status, feature, emotion_spiral
Settings: place, time, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal
}
