% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dataset_keys}
\alias{dataset_keys}
\title{Get the list of dataset keys}
\usage{
dataset_keys(dicts = get_dicts())
}
\arguments{
\item{dicts}{list of dictionary objects}
}
\value{
list of dictionary keys
}
\description{
Returns the names of available datasets
}
\examples{
dataset_keys()
}
