% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{dict_subset}
\alias{dict_subset}
\title{dict_subset}
\usage{
dict_subset(type = NA, component = NA, dicts = get_dicts())
}
\arguments{
\item{type}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{component}{string (\code{"identity"}, \code{"behavior"}, \code{"modifier"}, or \code{"setting"})}

\item{dicts}{list of dictionary objects}
}
\value{
list of dictionary keys
}
\description{
Dictionary subset. Extracts the keys of dictionaries, optionally subsetted by type (mean, sd, cov, individual) or component (identity, behavior, modifier, setting)
}
