% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egypt2015_behaviors_av_dict}
\alias{egypt2015_behaviors_av_dict}
\title{Egypt 2015 behaviors (gender = average)}
\format{
A data frame with 368 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{http://affectcontroltheory.org///egypt-dictionary-2015/}
}
\usage{
egypt2015_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
