% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egypt2015_individual}
\alias{egypt2015_individual}
\title{Individual data, Egypt, 2015}
\format{
A data frame with 124265 rows and 8 variables:
\describe{
\item{Condition}{}
\item{Gender}{respondent gender}
\item{UserID}{}
\item{component}{}
\item{term}{dictionary term}
\item{E}{evaluation rating}
\item{P}{potency rating}
\item{A}{activity rating}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
egypt2015_individual
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
