% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{eqn_subset}
\alias{eqn_subset}
\title{Equation subset}
\usage{
eqn_subset(eqns, gender = NA)
}
\arguments{
\item{eqns}{list of equation objects}

\item{gender}{string}
}
\value{
list of equation keys
}
\description{
Extracts the keys of equations, optionally subsetted by gender
}
\keyword{internal}
