% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\docType{class}
\name{equation-class}
\alias{equation-class}
\title{Equation class}
\value{
a new object
}
\description{
Equation class
}
\section{Slots}{

\describe{
\item{\code{key}}{character.}

\item{\code{genders}}{vector.}

\item{\code{filetype}}{character.}

\item{\code{source}}{character.}

\item{\code{description}}{character.}

\item{\code{citation}}{character.}
}}

