% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\title{All equation dataframes}
\format{
A nested data frame with 78 rows and 4 columns
\describe{
\item{key}{equation set name}
\item{equation_type}{type of equation}
\item{gender}{equation gender}
\item{df}{dataframe with equation coefficients}
...
}
}
\source{
Interact (May 2021)
}
\usage{
equations
}
\description{
This nested data frame provides information and coefficients for ACT equations.
Coefficient values are reported as given in Interact. If gender is listed as average,
this indicates that for the given study and subset, the values provided in
Interact for male and female are identical. To retrieve coefficients for particular
equation sets, I recommend using the \code{get_eqn()} function rather than working
with this data frame directly.
}
\details{
For more information about available equation files, call eqn_info()
}
\keyword{datasets}
