% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressive2002_behaviors_av_dict}
\alias{expressive2002_behaviors_av_dict}
\title{Expressive behaviors 2002 (gender = average)}
\format{
A data frame with 98 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/expressive_acts.htm}
}
\usage{
expressive2002_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
