% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaysex1980_behaviors_f_dict}
\alias{gaysex1980_behaviors_f_dict}
\title{Sexual behaviors 1980 (rated by gay men in San Francisco): subset of those with more negative views of unsafe sex (labeled "female" for interact purposes).}
\format{
A data frame with 14 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}

A data frame with 14 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/gay_sex.htm}

\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/gay_sex.htm}
}
\usage{
gaysex1980_behaviors_f_dict

gaysex1980_behaviors_f_dict
}
\description{
For more description and citation information, call dict_info()

For more description and citation information, call dict_info()
}
\keyword{datasets}
