% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germany1989_mods_f_dict}
\alias{germany1989_mods_f_dict}
\title{Germany 1989 modifiers (gender = female)}
\format{
A data frame with 68 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
"Interact (May 2021)"
}
\usage{
germany1989_mods_f_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
