% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_subset.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(dataset, component = "all", type = "all", gender = "all")
}
\arguments{
\item{dataset}{key or list of keys (use dict_info() for list)}

\item{component}{component or list of components (identity, behavior, modifier, setting; can abbreviate i, b, s, m)}

\item{type}{type or list of types (mean, sd, cov)}

\item{gender}{gender or list of genders (male, female, average; can abbreviate m, f, a or av)}
}
\value{
a dataframe containing the desired subset of the EPA dictionary summary data.
}
\description{
Helper function for filtering the dictionary summary data. This function allows the user to obtain subsets that include only specified datasets, components, genders, and columns.
If the specified combination does not exist, this function throws a warning to the user.
}
