% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{get_dicts}
\alias{get_dicts}
\title{Get dictionary information}
\usage{
get_dicts()
}
\value{
list of dictionary objects
}
\description{
Return metadata of available dictionaries as a list of dictionary objects
}
