% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_meta}
\alias{get_meta}
\title{Just passes along the metadata object from internal data (for readme purposes)}
\usage{
get_meta(type)
}
\arguments{
\item{type}{dict or eqn}
}
\value{
metatdata df
}
\description{
Just passes along the metadata object from internal data (for readme purposes)
}
\keyword{internal}
