% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household1994_behaviors_av_dict}
\alias{household1994_behaviors_av_dict}
\title{Household 1994 behaviors (gender = average)}
\format{
A data frame with 22 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/household.htm}
}
\usage{
household1994_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
