% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiana2003_identities_m_dict}
\alias{indiana2003_identities_m_dict}
\title{Indiana 2003 identities (gender = male)}
\format{
A data frame with 500 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
Interact (May 2021)
}
\usage{
indiana2003_identities_m_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
