% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual}
\alias{individual}
\title{Individual data}
\format{
A data frame with 674389 rows and 13 variables:
\describe{
\item{dataset}{the data collection effort the observation belongs to}
\item{context}{context in which dataset was collect (usually country)}
\item{year}{year dataset collected/published}
\item{userid}{individual identifier}
\item{gender}{respondent's gender identity}
\item{age}{respondent's age or age bracket}
\item{race1}{respondent's primary racial identification}
\item{race2}{respondent's secondary racial identification}
\item{term}{dictionary term}
\item{component}{the term's type}
\item{instcode}{institution code}
\item{E}{evaluation rating}
\item{P}{potency rating}
\item{A}{activity rating}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
individual
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
