% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{initialize,dictionary-method}
\alias{initialize,dictionary-method}
\title{Initializer method for dictionary class}
\usage{
\S4method{initialize}{dictionary}(
  .Object,
  key = NA_character_,
  components = c("identities", "behaviors", "mods"),
  types = c("mean"),
  genders = c("male", "female", "av"),
  filetype = ".dat",
  source = "Interact 2.1 beta (May 2021)",
  description = "unknown",
  citation = "unknown"
)
}
\arguments{
\item{.Object}{dictionary object (self)}

\item{key}{dictionary name}

\item{components}{available components}

\item{types}{available data types}

\item{genders}{available genders}

\item{filetype}{original source filetype}

\item{source}{where data came from}

\item{description}{description provided for data}

\item{citation}{citation provided for data}

\item{dictionary}{dictionary object}
}
\value{
a new dictionary object
}
\description{
sets defaults but allows them to be overridden by provided values
}
