% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{initialize,equation-method}
\alias{initialize,equation-method}
\title{Initializer for the equation class}
\usage{
\S4method{initialize}{equation}(
  .Object,
  key = NA_character_,
  genders = c("male", "female"),
  filetype = ".dat",
  source = "Interact 2.1 (May 2021)",
  description = "unknown",
  citation = "unknown"
)
}
\arguments{
\item{.Object}{equation object (self)}

\item{key}{equation set name}

\item{genders}{available genders}

\item{filetype}{original source filetype}

\item{source}{where data came from}

\item{description}{description provided for data}

\item{citation}{citation provided for data}

\item{equation}{equation object}
}
\description{
Initializer for the equation class
}
