% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internationaldomesticrelations1981_behaviors_av_dict}
\alias{internationaldomesticrelations1981_behaviors_av_dict}
\title{International and domestic relations 1981 behaviors (gender = average)}
\format{
A data frame with 238 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/international_relations.htm
}
\usage{
internationaldomesticrelations1981_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
