% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet1998_behaviors_av_dict}
\alias{internet1998_behaviors_av_dict}
\title{Internet 1998 behaviors (gender = average)}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/InternetCulture.htm
}
\usage{
internet1998_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
