% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japan19892002_behaviors_av_dict}
\alias{japan19892002_behaviors_av_dict}
\title{Japan 1989-2002 behaviors (gender = average)}
\format{
A data frame with 455 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
Interact (May 2021)
}
\usage{
japan19892002_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
