% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{just_once}
\alias{just_once}
\title{just_once}
\usage{
just_once(colnames, required)
}
\arguments{
\item{colnames}{vector the (bigger) list of column names}

\item{required}{vector the (smaller) list of required columns}
}
\value{
logical whether each element of required is in colnames once and only once
}
\description{
Tests that all elements of a required vector are in a vector of column names once and only once
}
