% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morocco2015_mods_av_dict}
\alias{morocco2015_mods_av_dict}
\title{Morocco 2015 modifiers (gender = average)}
\format{
A data frame with 232 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{http://affectcontroltheory.org///morocco-dictionary-2015/}
}
\usage{
morocco2015_mods_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
