% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc1978_mods_av_dict}
\alias{nc1978_mods_av_dict}
\title{North Carolina 1978 modifiers (gender = average)}
\format{
A data frame with 441 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
Interact (May 2021)
}
\usage{
nc1978_mods_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
