% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc1978_selfdir_f_eqn}
\alias{nc1978_selfdir_f_eqn}
\title{North Carolina 1978 self-directed action (gender = female)}
\format{
A data frame with 12 rows (terms) and 7 variables:
\describe{
\item{V1}{term}
\item{V2-VN}{equations}
...
}
}
\source{
Interact (May 2021)
}
\usage{
nc1978_selfdir_f_eqn
}
\description{
These values are reported as given in Interact. If gender is listed as average, this indicates that for the given study and subset, the values provided in Interact for male and female are identical.
}
\details{
For more information about available equation files, call eqn_info()
}
\keyword{datasets}
