% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontario1980_identities_av_dict}
\alias{ontario1980_identities_av_dict}
\title{Ontario 1980 identities (gender = average)}
\format{
A data frame with 809 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
Interact (May 2021)
}
\usage{
ontario1980_identities_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
