% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{politics2003_behaviors_m_dict}
\alias{politics2003_behaviors_m_dict}
\title{Politics 2003 behaviors (gender = male)}
\format{
A data frame with 21 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/politics.htm}
}
\usage{
politics2003_behaviors_m_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
