% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prisonersdilemma_behaviors_f_dict}
\alias{prisonersdilemma_behaviors_f_dict}
\title{Prisoner's dilemma behaviors (gender = female)}
\format{
A data frame with 2 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
\url{http://bayesact.ca/}
}
\usage{
prisonersdilemma_behaviors_f_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
