% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prisonersdilemmaCOV_behaviors_av_COV_dict}
\alias{prisonersdilemmaCOV_behaviors_av_COV_dict}
\title{Prisoner's dilemma behaviors with covariance information (gender = average)}
\format{
A data frame with 2 rows and 13 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{cov#}{covariance matrix value}
...
}
}
\source{
\url{http://bayesact.ca/}
}
\usage{
prisonersdilemmaCOV_behaviors_av_COV_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
