% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata}
\usage{
read_metadata(type)
}
\arguments{
\item{type}{dict or eqn}
}
\value{
dataframe
}
\description{
Reads in the metadata associated with equations and dictionaries available.
}
