% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{reformat}
\alias{reformat}
\title{Reformat data to bayesact or interact-friendly format depending on summary statistic desired.}
\usage{
reformat(data, program, stat = "mean", group = "none", groupvals = NA)
}
\arguments{
\item{data}{dataframe to reformat (already verified that column names and types are correct)}

\item{program}{bayesact or interact. Mean datasets look the same.}

\item{stat}{summary statistic desired (mean, sd, or cov; interact must be mean)}

\item{group}{name of grouping column (interact/mean only)}

\item{groupvals}{vector length 2; unique values in grouping column}
}
\value{
reformatted data frame.
Interact:
\itemize{
\item term
\item 6 EPA cols, repeated if no group, one set for each group if grouped
\item no institution codes
}

Bayesact:
\itemize{
\item term
\item 3 EPA cols
\item if mean: 3 more placeholder EPA cols
\item if sd: 3 sd cols
\item if cov: 9 cov cols
\item placeholder institution codes for all.
}
}
\description{
Assuming here that we've already checked that the necessary columns are present and named correctly.
This reorders them and drops the unnecessary ones. It also does the grouping if desired for interact files.
}
