% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_for_interact}
\alias{save_for_interact}
\title{Save for export to Interact}
\usage{
save_for_interact(
  data,
  type = "dict",
  filename = paste0(deparse(substitute(data)), ".txt")
)
}
\arguments{
\item{data}{dataframe to save}

\item{type}{string ("dict" or "eqn") indicating type of data}

\item{filename}{string the filepath at which to save (must end in .txt)}
}
\value{

}
\description{
Saves dictionary and equation files in a format that can be directly copy/pasted into Interact.
This may be useful for easily creating subsets of data to use in Interact or for importing dictionaries that are not already included.
}
