% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{saveit}
\alias{saveit}
\title{Save data files}
\usage{
saveit(..., name, type = NA)
}
\arguments{
\item{...}{data object}

\item{name}{desired object name (filename without type and extension)}

\item{type}{data type (dict or eqn)}
}
\description{
Adapted from https://stackoverflow.com/questions/21248065/r-rename-r-object-while-save-ing-it : see Marco Wirthlin's answer
}
