% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{standardize_inst_code}
\alias{standardize_inst_code}
\title{Checks whether an institution code is properly formatted and returns the code with the proper whitespace}
\usage{
standardize_inst_code(code)
}
\arguments{
\item{code}{character institution code}
}
\value{
character standardized institution code
}
\description{
Checks whether an institution code is properly formatted and returns the code with the proper whitespace
}
\keyword{internal}
