% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term_table_beh}
\alias{term_table_beh}
\title{Behavior terms}
\format{
A data frame with 1655 rows and 26 variables:
\describe{
\item{term}{behavior term}
\item{china1999}{dictionary}
\item{egypt2015}{dictionary}
\item{expressive2002}{dictionary}
\item{gaysex1980}{dictionary}
\item{germany1989}{dictionary}
\item{germany2007}{dictionary}
\item{household1994}{dictionary}
\item{indiana2003}{dictionary}
\item{internationaldomesticrelations1981}{dictionary}
\item{internet1998}{dictionary}
\item{japan19892002}{dictionary}
\item{morocco2015}{dictionary}
\item{nc1978}{dictionary}
\item{nireland1977}{dictionary}
\item{ontario1980}{dictionary}
\item{ontario2001}{dictionary}
\item{politics2003}{dictionary}
\item{prisonersdilemma}{dictionary}
\item{prisonersdilemmaCOV}{dictionary}
\item{texas1998}{dictionary}
\item{uga2015}{dictionary}
\item{uga2015bayesactsubset}{dictionary}
\item{usfullsurveyor2015}{dictionary}
\item{usmturk2015}{dictionary}
\item{usstudent2015}{dictionary}
...
}
}
\usage{
term_table_beh
}
\description{
Behavior terms contained in all publicly available dictionaries with indicators for which dictionary contains them.
Note that terms have been standardized between dictionaries to facilitate comparison, so terms here may differ from those in the original source data files in substantively unimportant ways (capitalization, spacing, spelling, etc).
There is one column per dictionary that includes this type of term. Whether or not a specific dictionary includes a particular term is indicated by 0/1 indicators.
}
\keyword{datasets}
