% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texas1998_identities_f_dict}
\alias{texas1998_identities_f_dict}
\title{Texas 1998 identities (gender = female)}
\format{
A data frame with 443 rows and 5 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{instcodes}{institutional codes for this term}
...
}
}
\source{
Interact (May 2021)
}
\usage{
texas1998_identities_f_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
