% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga2015bayesactsubset_identities_av_COV_dict}
\alias{uga2015bayesactsubset_identities_av_COV_dict}
\title{University of Georgia 2015 identities with covariance information (gender = average). Note that a slightly different subset of respondents was used to calculate these summary statistics than was used in the mean-only dataset (uga2015 in this package).}
\format{
A data frame with 929 rows and 13 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
\item{cov#}{covariance matrix value}
...
}
}
\source{
\url{http://bayesact.ca/}
}
\usage{
uga2015bayesactsubset_identities_av_COV_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
