% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usfullsurveyor2015_individual}
\alias{usfullsurveyor2015_individual}
\title{Individual data, full US surveyor sample, 2015}
\format{
A data frame with 75169 rows and 25 variables:
\describe{
\item{term}{dictionary term}
\item{term}{dictionary term}
\item{term}{dictionary term}
\item{term}{dictionary term}
\item{term}{dictionary term}
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
usfullsurveyor2015_individual
}
\description{
TODO: Finish updating
}
\details{
For more description and citation information, call dict_info()
}
\keyword{datasets}
