% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usmturk2015_behaviors_av_dict}
\alias{usmturk2015_behaviors_av_dict}
\title{US Amazon Turk 2015 behaviors (gender = average)}
\format{
A data frame with 853 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-online-dictionary-2015/}
}
\usage{
usmturk2015_behaviors_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
