% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usstudent2015_individual}
\alias{usstudent2015_individual}
\title{Individual data, US student sample, 2015}
\format{
A data frame with 157176 rows and 31 variables:
\describe{
\item{Condition}{}
\item{Gender}{respondent gender}
\item{UserID}{}
\item{Set}{}
\item{age}{respondent age}
\item{business}{}
\item{component}{}
\item{daded}{father's level of education}
\item{dademp}{father's employment status}
\item{grade}{}
\item{hispanic}{}
\item{home}{}
\item{home2}{}
\item{language}{}
\item{livedprior}{}
\item{momed}{mother's level of education}
\item{momemp}{mother's employment status}
\item{otherrelig}{}
\item{parentinc}{}
\item{parentmarital}{}
\item{proportionus}{}
\item{race1}{}
\item{race2}{}
\item{religion}{}
\item{relimp}{religious importance}
\item{services}{}
\item{usborn}{}
\item{term}{dictionary term}
\item{E}{evaluation rating}
\item{P}{potency rating}
\item{A}{activity rating}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
usstudent2015_individual
}
\description{
Combination of UGA and Duke samples
}
\details{
For more description and citation information, call dict_info()
}
\keyword{datasets}
