% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usstudent2015_mods_av_dict}
\alias{usstudent2015_mods_av_dict}
\title{US 2015 student sample (UGA and Duke) modifiers (gender = average)}
\format{
A data frame with 660 rows and 4 variables:
\describe{
\item{term}{dictionary term}
\item{E}{evaluation mean value}
\item{P}{potency mean value}
\item{A}{activity mean value}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
usstudent2015_mods_av_dict
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
