% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{check_stat}
\alias{check_stat}
\title{Valid stat check}
\usage{
check_stat(stat)
}
\arguments{
\item{stat}{string}
}
\value{
TRUE if available; throws error otherwise
}
\description{
Checks provided stat(s) to see if they are valid designators. Takes certain abbreviations/alternate spellings.
Accepted values are:
mean, m
sd, standard deviation, s
cov, covar, covariance, c
}
