% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dataset_keys}
\alias{dataset_keys}
\title{dict_subset}
\usage{
dataset_keys(dicts = get_dicts())
}
\arguments{
\item{dicts}{list of dictionary objects}
}
\value{
list of dictionary keys
}
\description{
Dictionary subset. Extracts the keys of dictionaries, optionally subsetted by stat (mean, sd, cov, individual) or component (identity, behavior, modifier, setting)
}
