% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dict_info}
\alias{dict_info}
\title{Print dictionary information}
\usage{
dict_info(name = NA)
}
\arguments{
\item{name}{string}
}
\description{
Print metadata for the dictionary name specified or for all available dictionaries (if name is unspecified)
}
\examples{
dict_info()
dict_info("china1999")
}
