% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dict_info}
\alias{dict_info}
\title{dict_info}
\usage{
dict_info(name = NA)
}
\arguments{
\item{name}{string}
}
\description{
Print dictionary info. Print metadata for the dictionary name specified or for all available dictionaries (if name is unspecified)
}
