% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dict_info}
\alias{dict_info}
\title{Print dictionary info}
\usage{
dict_info(name = NA)
}
\arguments{
\item{name}{string}
}
\value{

}
\description{
Print metadata for the dictionary name specified or for all available dictionaries (if name is unspecified)
}
\examples{
dict_info('nc1978')
}
