% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{dict_subset}
\alias{dict_subset}
\title{Dictionary subset}
\usage{
dict_subset(dicts, type = NA, component = NA)
}
\arguments{
\item{dicts}{list of dictionary objects}

\item{type}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{component}{string (\code{"identities"}, \code{"behaviors"}, \code{"mods"}, or \code{"settings"})}
}
\value{
list of dictionary keys
}
\description{
Extracts the keys of dictionaries, optionally subsetted by type (mean, SD, COV) or component (identities, behaviors, mods, settings)
}
\keyword{internal}
