% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{class}
\name{dictionary-class}
\alias{dictionary-class}
\alias{dictionary}
\title{Dictionary}
\value{
a dictionary object
}
\description{
Dictionary
}
\section{Slots}{

\describe{
\item{\code{key}}{character. The dictionary name and year.}

\item{\code{components}}{vector. What types of terms are included (identities, behaviors, mods, settings)}

\item{\code{types}}{vector. What type of data is available (mean, SD, COV)}

\item{\code{genders}}{vector. What genders are available (male, female, av)}

\item{\code{filetype}}{character. Original source file extension.}

\item{\code{source}}{character. Where original data came from.}

\item{\code{description}}{character. Description provided by the source for the dataset.}

\item{\code{citation}}{character. Citation information for the dataset.}

\item{\code{notes}}{character. Any notes for the dataset.}
}}

