% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dukestudent2015_individual}
\alias{dukestudent2015_individual}
\title{Individual data, Duke student sample, 2015}
\format{
A data frame with 21481 rows and 19 variables:
\describe{
\item{Condition}{}
\item{Gender}{respondent gender}
\item{UserID}{}
\item{age}{respondent age}
\item{component}{}
\item{daded}{father's level of education}
\item{dademp}{father's employment status}
\item{grade}{}
\item{home}{}
\item{home2}{}
\item{momed}{mother's level of education}
\item{momemp}{mother's employment status}
\item{race1}{}
\item{race2}{}
\item{relimp}{religious importance}
\item{term}{dictionary term}
\item{E}{evaluation rating}
\item{P}{potency rating}
\item{A}{activity rating}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
dukestudent2015_individual
}
\description{
For more description and citation information, call dict_info()
}
\keyword{datasets}
