% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_processing.R
\name{epa_summary}
\alias{epa_summary}
\title{epa_summary}
\usage{
epa_summary(data)
}
\arguments{
\item{data}{individual level data}
}
\value{
a summary dataset with one row per term. Includes the evaluation, potency, and activity mean, standard deviation, and variance-covariance matrix entries for each term.
}
\description{
Calculates the mean, standard deviation, and variance-covariance matrix for each term in one of the included individual datasets. This is useful when a user wants summary EPA information for a subset of respondents, for example, when comparing cultural meaning across groups.
}
