% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{eqn_info}
\alias{eqn_info}
\title{eqn_info}
\usage{
eqn_info(name = NA)
}
\arguments{
\item{name}{string}
}
\description{
Print equation metadata. Print metadata for requested equation name or (if name is not specified) for all available equations
}
