% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{eqn_info}
\alias{eqn_info}
\title{Print equation metadata}
\usage{
eqn_info(name = NA)
}
\arguments{
\item{name}{string}
}
\value{

}
\description{
Print metadata for requested equation name or (if name is not specified) for all available equations
}
\examples{
eqn_info('nc1978')
}
