% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{eqn_subset}
\alias{eqn_subset}
\title{eqn_subset}
\usage{
eqn_subset(eqns, gendercomponent = NA)
}
\arguments{
\item{eqns}{list of equation objects}

\item{gendercomponent}{string}
}
\value{
list of equation keys
}
\description{
Equation subset. Extracts the keys of equations, optionally subsetted by gender
}
