% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\docType{class}
\name{equation-class}
\alias{equation-class}
\title{Equation}
\value{
an equation object
}
\description{
Equation
}
\section{Slots}{

\describe{
\item{\code{key}}{character. The equation name and year}

\item{\code{gendercomponents}}{vector.}

\item{\code{filetype}}{character.}

\item{\code{source}}{character.}

\item{\code{description}}{character.}

\item{\code{citation}}{character.}
}}

