% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\title{All equation dataframes}
\format{
A data frame with 8 rows (terms) and 4 variables:
\describe{
\item{key}{equation set name}
\item{equation_type}{type of equation}
\item{gender}{equation gender}
\item{df}{dataframe with equation coefficients}
...
}
}
\source{
Interact (May 2021)
}
\usage{
equations
}
\description{
These values are reported as given in Interact. If gender is listed as average, this indicates that for the given study and subset, the values provided in Interact for male and female are identical.
}
\details{
For more information about available equation files, call eqn_info()
}
\keyword{datasets}
