% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dicts}
\alias{get_dicts}
\title{get_dicts}
\usage{
get_dicts()
}
\value{
list of dictionary objects
}
\description{
Get dictionary information. Return metadata of available dictionaries as a list of dictionary objects
}
