% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{get_eqn}
\alias{get_eqn}
\title{Extract single equation data frame}
\usage{
get_eqn(key, equation_type = "impressionabo", gender = "average")
}
\arguments{
\item{key}{the name of the equation set, Call eqn_info() or see package readme for options.}

\item{equation_type}{the type of the equation. Options: emotionid, impressionabo, impressionabos, selfdir, traitid}

\item{gender}{gender. Average, female, or male.}
}
\value{
equation dataframe
}
\description{
Extract single equation data frame
}
