% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{get_eqns}
\alias{get_eqns}
\title{get_eqns}
\usage{
get_eqns()
}
\value{
list of equation objects
}
\description{
Get equation information. Return metadata for all available equations as a list of equation objects
}
