% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual}
\alias{individual}
\title{Individual data}
\format{
A data frame with 794449 rows and 16 variables:
\describe{
\item{dataset}{the data collection effort the observation belongs to}
\item{context}{context in which dataset was collect (usually country)}
\item{year}{year dataset collected/published}
\item{userid}{individual identifier}
\item{gender}{respondent's gender identity. Male/female except in
occupations data sets, where genderqueer/nonbinary and other are options as well.}
\item{age}{respondent's age or age bracket}
\item{raceeth}{respondent's racial/ethnic identification (relevant to artifactmods2022, humanvalues2022, products2022, occs2019, occs2020, occs2021)}
\item{race1}{respondent's primary racial identification (relevant to dukecommunity2015, dukestudent2015, uga2015, usmturk2015, usstudent2015, usfullsurveyor2015)}
\item{race2}{respondent's secondary racial identification (relevant to dukecommunity2015, dukestudent2015, uga2015, usmturk2015, usstudent2015, usfullsurveyor2015)}
\item{hisp}{whether a respondent identifies as Hispanic (relevant to uga2015, mturk2015, usstudent2015, usfullsurveyor2015)}
\item{term}{dictionary term}
\item{component}{the term's type}
\item{instcodes}{institution code}
\item{E}{evaluation rating}
\item{P}{potency rating}
\item{A}{activity rating}
...
}
}
\source{
\url{http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/}
}
\usage{
individual
}
\description{
This data frame contains respondent-level data from a number of recent data collections.
The public release version of these data has been trimmed in order to preserve respondent
privacy. Covariates retained, where available, are gender identification, racial/ethnic
identification, and age. Full versions of these data may be available from the authors.
Call \code{dict_info()} for citation information.
}
\details{
Please note that covariates \emph{may not be comparable between data sets}. Different
data collections provided different response options for these questions. Only very minimal
recoding has been done for the version included here. Racial and ethnic identification
variables show the most differences between data sets. If you wish to use these variables
for research, you should contact the data set authors in order to obtain questionnaire information.
}
\keyword{datasets}
