% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{initialize,dictionary-method}
\alias{initialize,dictionary-method}
\title{Initializer method for dictionary class}
\usage{
\S4method{initialize}{dictionary}(
  .Object,
  key = NA_character_,
  context = NA_character_,
  year = NA_character_,
  components = c("identity", "behavior", "modifier"),
  stats = c("mean"),
  genders = c("male", "female", "average"),
  individual = FALSE,
  filetype = ".dat",
  source = "Interact 2.1 beta (May 2021)",
  description = "unknown",
  citation = "unknown",
  notes = "none"
)
}
\arguments{
\item{.Object}{dictionary object (self)}

\item{key}{dictionary name}

\item{context}{country or context collected in}

\item{year}{year collected}

\item{components}{available components}

\item{stats}{available summary statistics}

\item{genders}{available genders}

\item{individual}{whether individual data is available}

\item{filetype}{original source filetype}

\item{source}{where data came from}

\item{description}{description provided for data}

\item{citation}{citation provided for data}

\item{notes}{any notes for the dataset}

\item{dictionary}{dictionary object}
}
\value{
a new dictionary object
}
\description{
sets defaults but allows them to be overridden by provided values
}
\keyword{internal}
