% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{initialize,equation-method}
\alias{initialize,equation-method}
\title{Initializer for the equation class}
\usage{
\S4method{initialize}{equation}(
  .Object,
  key = NA_character_,
  gendercomponents = c("impressionabo_f", "impressionabo_m", "impressionabos_f",
    "impressionabos_m", "selfdir_f", "selfdir_m", "traitid_f", "traitid_m",
    "emotionid_f", "emotionid_m"),
  filetype = ".dat",
  source = "Interact 2.1 (May 2021)",
  description = "unknown",
  citation = "unknown",
  notes = "none"
)
}
\arguments{
\item{.Object}{equation object (self)}

\item{key}{equation set name}

\item{gendercomponents}{component genders}

\item{filetype}{original source filetype}

\item{source}{where data came from}

\item{description}{description provided for data}

\item{citation}{citation provided for data}

\item{notes}{any notes provided for the dataset}

\item{equation}{equation object}
}
\value{
a new equation object
}
\description{
sets defaults but allows them to be overridden by provided values
}
\keyword{internal}
