% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{reformat}
\alias{reformat}
\title{Reformat data to bayesact or interact-friendly format depending on summary statistic desired.}
\usage{
reformat(data, program, stat = "mean", group = "none", groupvals = NA)
}
\arguments{
\item{data}{dataframe to reformat (already verified that column names and types are correct)}

\item{program}{bayesact or interact. Mean datasets look the same.}

\item{stat}{summary statistic desired (mean, sd, or cov; interact must be mean)}

\item{group}{name of grouping column (interact/mean only)}

\item{groupvals}{vector length 2; unique values in grouping column}
}
\value{
reformatted data frame.
}
\description{
Assuming here that we've already checked that the necessary columns are present and named correctly.
This reorders them and drops the unnecessary ones. It also does the grouping if desired for interact files.
}
\keyword{-}
\keyword{3}
\keyword{6}
\keyword{9}
\keyword{Bayesact:}
\keyword{EPA}
\keyword{Interact:}
\keyword{all.}
\keyword{codes}
\keyword{cols}
\keyword{cols,}
\keyword{cov}
\keyword{cov:}
\keyword{each}
\keyword{for}
\keyword{group}
\keyword{group,}
\keyword{grouped}
\keyword{if}
\keyword{institution}
\keyword{internal}
\keyword{mean:}
\keyword{more}
\keyword{no}
\keyword{one}
\keyword{placeholder}
\keyword{repeated}
\keyword{sd}
\keyword{sd:}
\keyword{set}
\keyword{term}
