% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_for_interact}
\alias{save_for_interact}
\title{save_for_interact}
\usage{
save_for_interact(
  data,
  type = "dict",
  group = "none",
  filename = paste0(deparse(substitute(data)), ".txt"),
  savefile = TRUE
)
}
\arguments{
\item{data}{dataframe to save}

\item{type}{string ("dict" or "eqn") indicating type of data}

\item{group}{string indicating the name or numeric index of a column to group on. This column must contain only two values and each term must have one entry for each}

\item{filename}{string the filepath at which to save (must end in .txt)}

\item{savefile}{logical that denotes whether to actually save the file (FALSE is primarily useful for testing purposes). Default is TRUE.}
}
\value{
the dataframe that is written to file
}
\description{
Save for export to Interact
}
\details{
Saves dictionary and equation files in a format that can be directly copy/pasted into Interact.
This may be useful for easily creating subsets of data to use in Interact or for importing dictionaries that are not already included.
}
