% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{search_dict}
\alias{search_dict}
\title{search_dict}
\usage{
search_dict(expr, key = "all", component = "all", gender = "all")
}
\arguments{
\item{term}{A term or regular expression to search}

\item{dict}{The key of the dictionary (or list of multiple) to search in. Default is "all."}
}
\value{
FALSE if the provided term or expression is not in any provided dictionary. If it occurs at least once, returns the subset of the dictionary(s) of interest where the term matches the provided expression.
}
\description{
Dictionary term search
}
