% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term_table_set}
\alias{term_table_set}
\title{Setting terms}
\format{
A data frame with 503 rows and 8 variables:
\describe{
\item{term}{dictionary term}
...
}
}
\usage{
term_table_set
}
\description{
Setting terms contained in all publicly available dictionaries with indicators for which dictionary contains them.
Note that terms have been standardized between dictionaries to facilitate comparison, so terms here may differ from those in the original source data files in substantively unimportant ways (capitalization, spacing, spelling, etc).
There is one column per dictionary that includes this type of term. Whether or not a specific dictionary includes a particular term is indicated by 0/1 indicators.
}
\keyword{datasets}
