% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{this_dict}
\alias{this_dict}
\title{Get dictionary or equation object}
\usage{
this_dict(name)
}
\arguments{
\item{name}{string}
}
\value{
dictionary object
}
\description{
Extracts metadata object for a single dictionary of a given name.
Its main purpose is to allow for testing whether particular dictionaries
have particular elements.
}
\examples{
this_dict("uga2015")
}
