% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_eq_utils.R
\name{this_dict}
\alias{this_dict}
\title{Get dictionary or equation object}
\usage{
this_dict(name, class = "dictionary")
}
\arguments{
\item{name}{string}

\item{class}{string (\code{"equation"} or \code{"dictionary"})}
}
\value{
dictionary object
}
\description{
Extracts metadata object for a single dictionary or equation of a given name from the master list
}
